i added context processor see description why
I added a context processor function called inject_user() inside the create_app() function in __init__.py.
This function ensures that, if a user is logged in, the username is always available in the templates without explicitly passing it in each route.
This solves the problem where the navbar didn’t properly update to show whether a user was logged in or not.
i needed to import session and User for it to work